refactor: move sim connect conversion funcs
